Add snippets for filter list of each kind
